Move JSONtoCSV logic to utils file
